requirements/base.txt
django==4.0
python-decouple==3.8
requests==2.28.2
